refactor: 오탈자 및 함수명 수정
