+++ src/components/MovieCard.ts
@@ -7,7 +7,7 @@ class MovieCard {
   constructor(movieDate: MovieItem) {
     this.movieData = movieDate;
     this.createTemplate();
-    this.initEventHandelr();
+    this.setEvent();
   }
 
   get node(): HTMLElement {
@@ -54,7 +54,7 @@ class MovieCard {
       'https://user-images.githubusercontent.com/112997662/223046479-306cc6a7-7024-4616-b28e-be2f2878d2f0.png';
   }
 
-  initEventHandelr() {
+  setEvent() {
     const thumbnail = this._node.querySelector<HTMLImageElement>('#item-thumbnail');
     const thumbnailSkeleton = this._node.querySelector<HTMLDivElement>('.skeleton');
