+++ src/render/apiError.ts
@@ -1,4 +1,4 @@
-import moveList from '../components/MoiveList';
+import moveList from '../components/MovieList';
 
 export default () => {
   moveList.removeSkeleton();
